+++ src/subclone.rs
@@ -1,3 +1,4 @@
+use crate::genotype::NeutralMutationPoisson;
 use crate::{stemcell::StemCell, write2file, MAX_SUBCLONES};
 use anyhow::{ensure, Context};
 use rand::seq::IteratorRandom;
@@ -16,10 +17,11 @@ pub struct Distributions {
     pub bern_asymmetric: Bernoulli,
     pub u: f64,
     no_asymmetric_division: bool,
+    pub neutral_poisson: NeutralMutationPoisson,
 }
 
 impl Distributions {
-    pub fn new(p_asymmetric: f64, u: f64, verbosity: u8) -> Self {
+    pub fn new(p_asymmetric: f64, u: f64, background: f32, division: f32, verbosity: u8) -> Self {
         if verbosity > 1 {
             println!(
                 "creating distributions with parameters asymmetric: {}, p_fitness: {}",
@@ -32,6 +34,7 @@ impl Distributions {
             bern_asymmetric: Bernoulli::new(p_asymmetric).expect("Invalid p: p<0 or p>1"),
             u,
             no_asymmetric_division,
+            neutral_poisson: NeutralMutationPoisson::new(division, background).unwrap(),
         }
     }
 
@@ -50,6 +53,7 @@ impl Default for Distributions {
             bern_asymmetric: Bernoulli::new(0.).unwrap(),
             u: f64::default(),
             no_asymmetric_division: true,
+            neutral_poisson: NeutralMutationPoisson::default(),
         }
     }
 }
@@ -475,7 +479,7 @@ mod tests {
     #[quickcheck]
     fn division_no_new_clone(seed: u64, cells_present: u8) -> bool {
         let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let distr = Distributions::new(1., 0f64, 0);
+        let distr = Distributions::new(1., 0f64, 1.1, 1.1, 0);
 
         let cells = vec![StemCell::new(); cells_present as usize];
         let before_assignment = cells.len();
@@ -489,7 +493,7 @@ mod tests {
     #[quickcheck]
     fn division_new_clone(seed: u64, cells_present: u8) -> bool {
         let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let distr = Distributions::new(1., 1f64, 0);
+        let distr = Distributions::new(1., 1f64, 1.1, 1.1, 0);
 
         let cells = vec![StemCell::new(); cells_present as usize];
         let before_assignment = cells.len();
@@ -504,7 +508,7 @@ mod tests {
     #[should_panic]
     fn assign_all_clones_occupied() {
         let mut rng = ChaCha8Rng::seed_from_u64(26);
-        let distr = Distributions::new(1., 1f64, 0);
+        let distr = Distributions::new(1., 1f64, 1.1, 1.1, 0);
 
         let mut subclones = SubClones::default();
         let cell = StemCell::new();
@@ -515,18 +519,18 @@ mod tests {
     #[should_panic]
     #[test]
     fn new_distribution_wrong_p_test() {
-        Distributions::new(1., f64::NAN, 0);
+        Distributions::new(1., f64::NAN, 1.1, 1.1, 0);
     }
 
     #[should_panic]
     #[test]
     fn new_distribution_wrong_p_inf_test() {
-        Distributions::new(1., f64::INFINITY, 0);
+        Distributions::new(1., f64::INFINITY, 1.1, 1.1, 0);
     }
 
     #[should_panic]
     #[test]
     fn new_distribution_wrong_p_neg_test() {
-        Distributions::new(1., -0.9, 0);
+        Distributions::new(1., -0.9, 1.1, 1.1, 0);
     }
 }